+++ src/Abstractions/NexusMods.Abstractions.Installers/NexusMods.Abstractions.Installers.csproj
@@ -15,21 +15,6 @@
     </ItemGroup>
 
     <ItemGroup>
-      <Compile Update="ILibraryFileInstaller.cs">
-        <DependentUpon>ILibraryItemInstaller.cs</DependentUpon>
-      </Compile>
-      <Compile Update="ILibraryArchiveInstaller.cs">
-        <DependentUpon>ILibraryItemInstaller.cs</DependentUpon>
-      </Compile>
-      <Compile Update="ALibraryItemInstaller.cs">
-        <DependentUpon>ILibraryItemInstaller.cs</DependentUpon>
-      </Compile>
-      <Compile Update="ALibraryFileInstaller.cs">
-        <DependentUpon>ILibraryFileInstaller.cs</DependentUpon>
-      </Compile>
-      <Compile Update="ALibraryArchiveInstaller.cs">
-        <DependentUpon>ILibraryArchiveInstaller.cs</DependentUpon>
-      </Compile>
       <Compile Update="AModInstaller.cs">
         <DependentUpon>IModInstaller.cs</DependentUpon>
       </Compile>
