+++ src/Abstractions/NexusMods.Abstractions.Library.Installers/ILibraryItemInstaller.cs
@@ -3,24 +3,24 @@ using NexusMods.Abstractions.Library.Models;
 using NexusMods.Abstractions.Loadouts;
 using NexusMods.MnemonicDB.Abstractions;
 
-namespace NexusMods.Abstractions.Installers;
+namespace NexusMods.Abstractions.Library.Installers;
 
 /// <summary>
-/// Variant of <see cref="ILibraryItemInstaller"/> for <see cref="LibraryFile"/>.
+/// Turns <see cref="LibraryItem"/> into <see cref="LoadoutItem"/>.
 /// </summary>
 [PublicAPI]
-public interface ILibraryFileInstaller : ILibraryItemInstaller
+public interface ILibraryItemInstaller
 {
     /// <summary>
-    /// Checks whether the provided library file is supported by this installer.
+    /// Checks whether the provided library item is supported by this installer.
     /// </summary>
-    ValueTask<bool> IsSupportedAsync(LibraryFile.ReadOnly libraryFile, CancellationToken cancellationToken);
+    ValueTask<bool> IsSupportedAsync(LibraryItem.ReadOnly libraryItem, CancellationToken cancellationToken);
 
     /// <summary>
     /// Executes the installer.
     /// </summary>
     ValueTask<LoadoutItem.New[]> ExecuteAsync(
-        LibraryFile.ReadOnly libraryFile,
+        LibraryItem.ReadOnly libraryItem,
         ITransaction transaction,
         Loadout.ReadOnly loadout,
         CancellationToken cancellationToken);
