+++ src/NexusMods.Library/InstallLoadoutItemJob.cs
@@ -1,6 +1,6 @@
 using DynamicData.Kernel;
-using NexusMods.Abstractions.Installers;
 using NexusMods.Abstractions.Jobs;
+using NexusMods.Abstractions.Library.Installers;
 using NexusMods.Abstractions.Library.Models;
 using NexusMods.Abstractions.Loadouts;
 using NexusMods.MnemonicDB.Abstractions;
